move codegen into shared_types
