+++ cli/Cargo.toml
@@ -12,5 +12,4 @@ chrono = "0.4.22"
 clap = { version = "4.0.18", features = ["derive"] }
 serde = { version = "1.0.147", features = ["derive"] }
 shared = { path = "../shared" }
-shared_types = { path = "../shared_types" }
 surf = "2.3.2"
